backend/requirements.txt
pyyaml
uritemplate==4.1.1
celery==5.2.2
Django==3.2.23
redis==3.5.3
psycopg2-binary
djantic==0.7.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
django-cors-headers==3.13.0
joblib==1.3.2
easyDataverse==0.4.1
